pyproject.toml
[tool.poetry]
name = "DESops"
version = "20.3.d0"
description = ""
authors = ["Jack Weitze <jweitze@umich.edu>", "Romulo Meira Goes <r.meiragoes@gmail.com>", "Shoma Matsui <smatsui@umich.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
python-igraph = "^0.8.0"
pycairo = "^1.19.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
seed-isort-config = "^2.1.0"
pre-commit = "^2.2.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
